+++ modules/client/src/erc20/internal/client/methods.ts
@@ -6,6 +6,7 @@ import {
   InvalidProposalIdError,
   NoProviderError,
   Random,
+  FailedToPinIPFS
 } from "@aragon/sdk-common";
 import { formatEther } from "@ethersproject/units";
 import {
@@ -86,7 +87,7 @@ export class ClientErc20Methods extends ClientCore
       // TODO: Compute the cid instead of uploading to the cluster
       cid = await this.ipfs.add(JSON.stringify(params.metadata));
     } catch {
-      throw new Error("Could not pin the metadata on IPFS");
+      throw new FailedToPinIPFS()
     }
 
     const startTimestamp = params.startDate?.getTime() || 0;