+++ modules/client/test/integration/client-addressList/methods.test.ts
@@ -90,7 +90,6 @@ describe("Client Address List", () => {
         },
         actions: [action],
         creatorVote: VoteValues.YES,
-        startDate: new Date(),
         executeOnPass: false,
       };